Remove plugins set to auto on
We had some hardcoded plugins using WP cli set to auto turn on when the
site gets built. Unfortunately we don't need some of these plugins and
we don't want them to be auto turned on so I've removed these except for
s3 which I think we'll continue to need to be on.